vzctl snapshot-switch: don't remove dump file
Usually, after successful restore from a dumpfile, the
dumpfile is removed since the CT is started, the file system
is changed and the dumpfile is no longer in a state consistent
with file system. This is the way it works since commit aec4998.

There should be one exception though. When ploop snapshots
are used, we do have consistent filesystem state and can
return to it many times, so we should not delete dump file
after successful restore.

+++ src/lib/hooks_vz.c
@@ -856,9 +856,9 @@ err:
 		close(dump_fd);
 	if (!ret) {
 		logger(0, 0, "Restoring completed successfully");
-		if (cmd == CMD_RESTORE)
-			if (dumpfile)
-				unlink(dumpfile);
+		if ((cmd == CMD_RESTORE) && (dumpfile) &&
+				!(skip & SKIP_DUMPFILE_UNLINK))
+			unlink(dumpfile);
 	}
 	return ret;
 }
